+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 4.37.0 (Unreleased)
+
+### Added
+- Support for oci_metering_computation_custom_table
+- Support for ExaCS Customer Contact
+- Support for Email Authentication: DKIM API
+
+### Fixed
+- Update issue in `oci_database_autonomous_database` - Issue [1417](https://github.com/terraform-providers/terraform-provider-oci/issues/1417)
+
 ## 4.36.0 (July 21, 2021)
 
 ### Added